Refactored DaoTests, implemented TestEntityFactory
Refactored Find, FindAll and Save tests of ProjectService

+++ project-portfolio-management-service/src/test/java/com/mbor/dao/UserDaoTest.java
@@ -1,7 +1,10 @@
 package com.mbor.dao;
 
+import com.mbor.configuration.TestConfiguration;
 import com.mbor.domain.Consultant;
+import com.mbor.domain.Employee;
 import com.mbor.domain.security.User;
+import com.mbor.entityFactory.TestEntityFactory;
 import com.mbor.spring.ServiceConfiguration;
 import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
@@ -15,12 +18,13 @@ import org.springframework.transaction.annotation.Transactional;
 import javax.persistence.EntityManager;
 import javax.persistence.EntityManagerFactory;
 import javax.persistence.EntityTransaction;
+import java.util.Optional;
 import java.util.Random;
 
 import static org.junit.jupiter.api.Assertions.*;
 
 @ExtendWith({SpringExtension.class})
-@ContextConfiguration(classes = ServiceConfiguration.class)
+@ContextConfiguration(classes = {ServiceConfiguration.class, TestConfiguration.class})
 @ActiveProfiles("test")
 @Transactional
 class UserDaoTest {
@@ -34,7 +38,7 @@ class UserDaoTest {
     private IUserDao userDao;
 
     @BeforeAll
-    static void init(@Autowired EntityManagerFactory entityManagerFactory) {
+    static void init(@Autowired EntityManagerFactory entityManagerFactory, @Autowired TestEntityFactory testEntityFactory) {
         EntityManager entityManager = entityManagerFactory.createEntityManager();
         EntityTransaction transaction =  entityManager.getTransaction();
         Random random = new Random();
@@ -46,17 +50,26 @@ class UserDaoTest {
         transaction.commit();
     }
 
+
     @Test
     public void saveThenSuccess() {
         User user = new User();
-        user.setEmployee(employeeDao.find(employeeID).get());
+        Optional<Employee> employeeOptional = employeeDao.find(employeeID);
+        if(!employeeOptional.isPresent()){
+            fail();
+        }
+        user.setEmployee(employeeOptional.get());
         assertNotNull(userDao.save(user).get());
     }
 
     @Test
     public void saveAndFindThenSuccess() {
         User user = new User();
-        user.setEmployee(employeeDao.find(employeeID).get());
+        Optional<Employee> employeeOptional = employeeDao.find(employeeID);
+        if(!employeeOptional.isPresent()){
+            fail();
+        }
+        user.setEmployee(employeeOptional.get());
         String username = userDao.save(user).get().getEmployee().getUserName();
         assertNotNull(userDao.findByUsername(username).get());
     }
@@ -64,7 +77,10 @@ class UserDaoTest {
     @Test
     public void deleteThenSuccess(){
         User user = new User();
-        user.setEmployee(employeeDao.find(employeeID).get());
+        Optional<Employee> employeeOptional = employeeDao.find(employeeID);
+        if(!employeeOptional.isPresent()){
+            fail();
+        }
         Long userId = userDao.save(user).get().getId();
         userDao.delete(userId);
         assertFalse(userDao.find(userId).isPresent());
